Exception handling - delete - atualização

+++ src/main/java/com/br/estudoCrippa/course/services/UserService.java
@@ -35,9 +35,11 @@ public class UserService {
 	
 	public void delete(Long id) {
 		try {
-			repository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
+			 if (repository.existsById(id)) {
+		            repository.deleteById(id);			
+		        } else {				
+		            throw new ResourceNotFoundException(id);			
+		        }
 		} catch (DataIntegrityViolationException e) {
 			throw new DatabaseException(e.getMessage());
 		}
